Added OpenAI TTS, camera timer, and phrases from the openai api
Open AI is used now, we generate text from it based on it. We give the user a phrase off the bat and then wait 5 minutes and throw it another phrase.

I have tried pyttsx3 which didn't work after the first saying. This is the best and cheapest method, i may come back to this later. I settled on pygame because it's very simple... Like... really simple, all I do is save the generated audio from OpenAI as a mp3 and just have pygame run the file and delete it when it's done.

Timer is added on the camera just because.

Overall this is working fine now... Just need to get the model going and have it read what it sees from the camera.

I've worked on this for at least an hour or two the past 3 days and the problem was the TTS and the damn solution was pygame......... I gotta thank AI for that.